refactor: RandomNumber 클래스 domain -> util 로 이동
