+++ src/main/java/step5/util/RandomNumberGenerator.java
@@ -1,13 +1,13 @@
-package step5.domain;
+package step5.util;
 
 import java.util.Random;
 
-public class RandomNumber {
+public class RandomNumberGenerator {
 
     private static final int RANDOM_NUMBER_RANGE = 10;
     private static final Random random = new Random();
 
-    public static int getRandomNumber() {
+    public static int generate() {
         return random.nextInt(RANDOM_NUMBER_RANGE);
     }
 }
